Several fixes and refactors
* Update react-native-screens to fix iOS title bug
* Refactor titles to be rendered instantly
* Move PersonTile to Tiles
* Fix date formatting shifting zones
