+++ package.json
@@ -20,6 +20,13 @@
   "jest": {
     "preset": "jest-expo"
   },
+  "expo": {
+    "install": {
+      "exclude": [
+        "react-native-screens"
+      ]
+    }
+  },
   "dependencies": {
     "@expo/vector-icons": "^14.0.4",
     "@react-navigation/native": "^6.1.18",
@@ -50,7 +57,7 @@
     "react-native-pressable-scale": "^2.1.0",
     "react-native-reanimated": "~3.10.1",
     "react-native-safe-area-context": "4.10.5",
-    "react-native-screens": "3.31.1",
+    "react-native-screens": "3.32.0",
     "react-native-svg": "15.2.0",
     "react-native-track-player": "^4.1.1",
     "react-native-web": "~0.19.13",
